src/backend/servico_analise_credito/requirements.txt
# Framework web e servidor ASGI
fastapi
uvicorn[standard]

# ORM e Migrações de Banco de Dados
sqlalchemy[asyncio]
alembic
asyncpg

# Validação de dados
pydantic

# Adicione esta linha para lidar com JWT
python-jose[cryptography]